pscom: Fix signature of pscom_malloc()
Switch from unsigned int to size_t for pscom_malloc().
(This commit just concerns the USE_PSCOM_MALLOC case.)

+++ lib/pscom/pscom_req.c
@@ -74,7 +74,7 @@ static
 struct list_head mallocs = LIST_HEAD_INIT(mallocs);
 
 
-void *pscom_malloc(unsigned int size)
+void *pscom_malloc(size_t size)
 {
 	void *ptr;
 	pscom_malloc_t *m;
